% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/details.R
\name{details}
\alias{details}
\title{create markdown details}
\usage{
details(text, summary = NULL, lang = "r")
}
\arguments{
\item{text}{character, text to put in details block}

\item{summary}{character, text to put in summary block, Default: NULL}

\item{lang}{character, language of block (for markdown highlighting) Default: 'r'}
}
\value{
character
}
\description{
Create details block for markdown with summary as optional.
}
