% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold.R
\name{foldstart}
\alias{foldstart}
\alias{foldend}
\title{Roxygen2 Documentation Folding}
\usage{
foldstart(title = NULL)

foldend()
}
\arguments{
\item{title}{Title of the folded section, Default: NULL}
}
\value{
character
}
\description{
Insert more information to function documentation
without overwhelming the user by using foldstart and foldend macros.
}
\details{
To enable this feature add to the  \code{DESCRIPTION} file
before rendering the \code{roxygen2}:

\strong{Roxygen: list(markdown = TRUE)}
}
\examples{

# With title

 "#' \foldstart{This is an example}
 
  #' DOCUMENTATION
 
  #' \foldend"


# With no title

 "#' \foldstart{}
 
  #' DOCUMENTATION
 
  #' \foldend"
 
}
\concept{fold}
