% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_details}
\alias{use_details}
\title{Use details macros}
\usage{
use_details(desc_path = "./DESCRIPTION")
}
\arguments{
\item{desc_path}{character, path to description file to check/edit, Default = './DESCRIPTION'}
}
\description{
Call this to import the details Rd macro into your package.
\itemize{
\item The RdMacros field of the DESCRIPTION file is updated so you can use the
\verb{\\\\foldstart\{\}} and \verb{\\\\foldend\{\}} macro in your documentation.
}

See the \href{https://yonicd.github.io/details/articles/documentation.html}{Package Documentation} vignette of the details package.
}
\details{
This will append three elements to the DESCRIPTION file
\itemize{
\item Suggests: details
\item RdMacros: details
\item Roxygen: list(markdown = TRUE)
}
}
\seealso{
\link[=details]{details}
}
