% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/details.R
\name{details}
\alias{details}
\title{Create HTML DOM Details}
\usage{
details(object, summary = NULL, tooltip = "Click to Expand",
  open = FALSE, lang = "r", device = FALSE, output = c("console",
  "clipr", "edit", "character"))
}
\arguments{
\item{object}{object, object to put in details block}

\item{summary}{character, text to put in summary block, Default: NULL}

\item{tooltip}{character, text for tooltip on the summary, Default: 'Click to Expand'}

\item{open}{logical, is the details open (TRUE) or closed (FALSE), Default: FALSE}

\item{lang}{character, language of block (for markdown highlighting) Default: 'r'}

\item{device}{logical, If TRUE then a device will be created internally to upload
the input into a png and invoke \link[knitr:upload_imgur]{upload_imgur} to place a URL
link pointing the output.}

\item{output}{character, where to output the file console (Default), clipboard or R file editor,
Default: c('console','clipr','file.edit','character')}
}
\value{
character
}
\description{
Create HTLM DOM Details block for Markdown documents
with summary as optional.
}
\details{
To remove summary or tooltip set them to NULL.
If the object is a file path, it will automatically it's lines will be read in internally.
}
