% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/details.R
\name{details}
\alias{details}
\title{create markdown details}
\usage{
details(text, summary = NULL, tooltip = "Click to Expand",
  open = FALSE, lang = "r", output = c("console", "clipr",
  "file.edit"))
}
\arguments{
\item{text}{character, text to put in details block}

\item{summary}{character, text to put in summary block, Default: NULL}

\item{tooltip}{character, text for tooltip on the summary, Default: 'Click to Expand'}

\item{open}{logical, is the details open (TRUE) or closed (FALSE), Default: FALSE}

\item{lang}{character, language of block (for markdown highlighting) Default: 'r'}
}
\value{
character
}
\description{
Create details block for markdown with summary as optional.
}
\details{
To remove summary or tooltip set them to NULL
}
