% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/details.R
\name{details}
\alias{details}
\title{create markdown details}
\usage{
details(text, summary = NULL, tooltip = "Click to Expand",
  open = FALSE, lang = "r", output = c("console", "clipr", "edit",
  "character"))
}
\arguments{
\item{text}{character, text to put in details block}

\item{summary}{character, text to put in summary block, Default: NULL}

\item{tooltip}{character, text for tooltip on the summary, Default: 'Click to Expand'}

\item{open}{logical, is the details open (TRUE) or closed (FALSE), Default: FALSE}

\item{lang}{character, language of block (for markdown highlighting) Default: 'r'}

\item{output}{character, where to output the file console (Default), clipboard or R file editor,
Default: c('console','clipr','file.edit','character')}
}
\value{
character
}
\description{
Create details block for markdown with summary as optional.
}
\details{
To remove summary or tooltip set them to NULL.
If the text is a file path, it will automatically it's lines will be read in internally.

\foldstart{More Details}
Here more information can be inserted without overwhelming the user.

To enable this
feature add \code{Roxygen: list(markdown = TRUE)} to the DESCRIPTION file before rendering
the roxygen2

you can use this feature by wrapping documentation with the macros

\code{\\foldstart{[SUMMARY TEXT]}}

\code{#' DOCUMENTATION}

\code{\\foldend}

\foldend
}
